requirements.txt
# Core ADK + A2A
google-adk[a2a]>=0.5.0

# LiteLLM — pinned to safe version (1.82.7 and 1.82.8 are compromised per security advisory)
# See: https://github.com/google/adk-python/issues/5005
litellm==1.82.6

# FastAPI backend
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# Static analysis
pefile>=2024.8.26
yara-python>=4.5.0
python-magic>=0.4.27

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
